+++ CHANGELOG.md
@@ -7,10 +7,12 @@ All significant changes to this project will be documented in this file.
 ### Breaking changes
 
 * `CountMinSketch` now has a type parameter for the count type. Possible values are `u8` to `u64` and `i8` to `i64`.
+* `HllUnion::get_result` is renamed to `HllUnion::to_sketch`.
 
 ### New features
 
 * `CountMinSketch` with unsigned values now supports `halve` and `decay` operations.
+* `CpcSketch` and `CpcUnion` is now available for cardinality estimation.
 
 ## v0.2.0 (2025-01-14)